requirements.txt
langchain
langchain-core

langchain-openai
openai
langchain-anthropic

langchain_google_genai
google-generativeai

langchain-huggingface
transformers
huggingface-hub

python-dotenv

numpy
scikit-learn
pandas
streamlit
pymupdf

pydantic
pydantic[email]



 